requirements/dev.txt
pre-commit
isort
black
gitchangelog
pytest-snapshot

ffmpeg-python
jsonschema
jinja2
pyaml
mergedeep

-e .
